fix: move component
